requirements.txt
streamlit
google-generativeai

#-e .
